Add a conversion sketch.
Use pyoracc to parse a given ATF file and manually construct
and print an TEI document containing the lines. This is about
me learning the output tree generated by the parser and thinking
about how to serialize it as xml.

When run on SIL-034.atf it produces something similar to the
manually converted SIL-034.tei and which passes validation.