Write out non-OraccObject text children as comments.
This lets us parse more of the files, although it adds considerable
noise to the output as pyoracc sometimes includes the entire tablet
again as an OraccObject which isn't an instance of OraccObject.

That's clearly a bug, but at least we produce valid xml again.
Or we would if the object serialization didn't include '--'.