travis: More explicit config.
Travis defaults don't know about pipenv, so we need
and explicit script.

pycodestyle needs to be pointed at a specific file when there's
now packaging infrastructure.

NB it would be nice to run `pipenv check` in periodic tests
to detect bitrot, but it seems to check all the packages
the travis venc has installed, so we get false positives.