Use format strings for the looped xml generation.
This is a built-in templating sugar added in Python 3.6. It's
equivalent to calling .format on the end of the string with
referenced variable from the local scope, and generally
much less noisy to read.

Also fix some cut-and-paste errors which were responsible
for including the wrong object type and contents when
a non-node child of the text object is encountered.