Escape line content.
Transliteration markup can contain sequences ilike `<missing>`
which are interpreted as xml tags or entities. Escape the
contents of the lines when writing them out so they're correctly
preserved.

Also write out an opening `<div>` tag before the xml comments
containing unhandled objects. This is necessary to balance
the unconditional `</div>` at the close of the loop.