Disable info messages.
These get in the way when dumping the output to a file for
serving by cdli-cts-server. Should replace them with logging
calls eventually.

Also, add an xml processing directive with the doctype and
charset encoding. This should make it more portable to older
tools, especially since we're using utf-8 encoding for the lines.