Add support for translations.
Handle translation objects in a second pass, writing them out
in a sibling tei:div to the edition.

This is a draft and needs refactoring. The parse tree from pyoracc
is a bit messy in that the translation is under the object as a
sibling of the tablet surfaces, but has those same surfaces again
as its own children, so there's an extra level of iteration.