Add atf2cts.py.
Wrapper around atf2tei which segments a multi-object atf file,
converts each segment to xml and then writes it out to a data
directory with __cts__.xml metadata files, according to the
required file hierarchy used by the Capitains standard.

Quite rough with hard-coded textgroup and hacky poking of
the urns in and out of the various documents.