Add classes to represent __cts__.xml index files.
This encapsulates the data needed for the Canonical Text
Services file hierarchy and makes it simple to set the
required data and serialize it to a string or file.

Note the xml representation is lazily constructed by
the `xml` propery getter. The fragments are short,
so performance from multiple reads shouldn't be an
issue. It could be memoized by hacking the object
member lookup to invalidate the cache, but that's
a lot more code. In typical use it's only called once.

Includes some simple unit tests.

Turns out `Element.find` doesn't search the current element,
so we have to check the root specially which makes the
tests a bit harder to read than I'd like.