Append lang to document filename.
This matches the lang suffix on the edition urn in the __cts__.xml
metadata file. I think this is incorrect though. Most texts are
so short it makes more sense to keep them in a single document,
so the language (which is different between the edition and the
translations) isn't helpful. It's also unclear how to distinguish
transliteration from normalization by lang tag.