Use line numbers from atf.
Previously we were generating sequential line numbers so the tei
labels were regular. But since the numbers are used for citation,
it's better to reflect the label as it occurs in the source atf.

This way they can be directly compared with other representations,
which is especially helpful after breaks or when translations aren't
provided for every line.