Start a tei module for serializing the XML text.
Add some classes to model the major pieces which can create and
ElementTree representation for serialization to a file.

The tests aren't very strong because I don't understand how to
make xpath pull out the elements which should be there. For
now it just creates a tree and verifies it serializes without
error.