Verify tei attribute serialization.
Fix issues with the xml serialization class tests.

A Document sets an xml namespace attribute, but a bare
Header does not. That may be incorrect, but it means
we need to namespace-qualify xpath queries when traversing
a Document, but not an unattached Header.

Also, xpath queries are separated by '/' not whitespace!