Add more header information to tei.Header.
Support the same header information we include in the current
atf2tei template.

It would be nice to make this generic with some kind of config
overlay added by each layer of conversion, but for now just
do what the template does. It's not time to make a TEI library
yet.