atf2cts: Return Success and Failure classes from convert.
The convert function spawned in parallel for handle each atf
block individually was still returning the success count and
a pair of failed input arrays it was updating when it was
part of a serial loop. This is a bit confusing.

Instead, return a class instance to mark success of failure.
This is a rust-ism as doesn't work that well in python since
the match statement to accumulate the failure counts based
on the type of the problematic input awkward.