atf2cts: Return a flag tuple from convert.
Instead of the fancy class, just return a flag tuple. Since boolean
values are numbers these can be accumulated directly into the caller's
counts.

This is still a bit confusing, but a lot less code than before,
which I think is an improvement.

By keeping just a count of failures, we also limit memory consumption.