atf2cts: Restore refsDecl markup.
CTS guidelines specify the header should contain patterns for
constructing xpath pointers based on the reference part of
the cts urn. I think Nautilus uses these to extract the
requested passages.

Since it's specific to CTS, the object is in the cts module and
it's only added by atf2cts, not atf2tei. However, it does need
a member on tei.Header to be included in serialization.