Append a project-language component to the edition urn.
We now pass the capitains_units.cts tests except for
Duplicate passages. This is perhaps confused by the
translation sections?

This is the recommended urn style. The language extension is
used to distinguish the primary text from translations, which
we don't really need since we're including the translation
in the same document.

Perseus also appends a version number, I think based on the
idea that the URNs should point to immutable contents. That's
not realistic for this project, so I've skipped that part.