Set a cdli project component to edition urns.
Recommendations are for the edition urn to have a component
reflecting the project producing the transcription. That
would be cdli in this case, assuming the converter is being
run under the cdli umbrella. This feels a little redundant
but better maches semantically with what Perseus does,
where an edition urn is something like

    urn:cts:greekLit:tlg0548.tlg001.perseus-grc2

It's just that in our case both the cts authority (greekLit above)
and the project name (perseus above) are the same.