atf2cts: Make generate a textgroup if none is given.
For testing we've been placing all the converted documents in
a `test` textgroup. That's fine when the caller wants to control
how the texts are grouped.

For mass conversion, duplicating the cdli code number for the object
as a separate text group is one way to address the CTS guidelines
requirement that there be a two-level file hierarchy, even though
we don't have traditional groupings the was greekLit and latinLit
do.

So, if no textgroup is passed in, use the generated one.