atf2tei: Propagate surface names from the parser.
Section objects may have an optional name, `Surface a` and so on.
We use the object name as the reference marker in the output TEI
it's best to propagate the name if present instead of labelling
all such divisions the same.

This helps with parsing the hammurapi text, reducing the number
of pass references MyCapytains generates to 8K instead of 1.8M.