atf2cts: Write separate files for each translation.
It makes sense from the point of view of cuneiform scholarship
to consider each tranliteration, along with the normalization
and translations, if any, as a single document. However, for
Capitains to be able to find specific passages through the
refsDecl patterns it needs each edition and translation in
a separate file with its own tei header.