cts: Remove object level refsDecl.
Don't include the object label (tablet, seal, etc.) in the declared
reference scheme. For the vast majority of records there's only
one top-level label, so it's redundant.

For records with multiple top-level labels, it would probably be
easier to write them out as separate files, so that part of the
citation is part of the work designation in the urn, rather
than the reference.