Convert rulings to notes as well.
Also add a unit test to verify extraction of the $-line.

The other option here is to use a <milestone rend="horizontal-rule">
or similar. Oracc uses <note> within transliteration, and a <p>
within translations. I think note is ultimately better here;
double and triple rulings are a single conceptual element in ATF,
and it would be hard to keep that with multiple <milestone> elements.